for #19 - I've resorted to wildcard include from rezolver to the source in rezolver.shared to sort this.
Cannot get Rezolver.Shared to work properly as a shared source package - either using the new nuget-as-a-target experience or via a nuspec.  I get as far as getting Rezolver to build with it, but intellisense is unaware that the shared types exist.  Not ideal.

In fact some of the package generation stuff in msbuild doesn't really appear to work.

Also - separate note - docfx project appears to be broken now.  Will investigate after getting all the rest of the projects converted.