trying to see whether I can get dynamic decoration working.
It works for one level of decoration (without any code changes); but not more than that - because we just replace any RezolvedTarget whose type matches the argument type.  You end up with only the outermost decorator kicking in.

I'm not even sure if this is a pattern we should support, but if there's a way to get it working without too much effort, it'd be good to do so!